+++ script-technical-analysis.py
@@ -1,8 +1,10 @@
 from models.PyCryptoBot import PyCryptoBot
 from models.Trading import TechnicalAnalysis
-#from models.TradingAccount import TradingAccount
-#from models.AppState import AppState
 
 app = PyCryptoBot()
 df = app.getHistoricalData(app.getMarket(), app.getGranularity())
+
+model = TechnicalAnalysis(df)
+model.addATR(14)
+df = model.getDataFrame()
 print (df)